+++ BE/src/interfaces/event.ts
